refactor(web): share command-palette open-state hook
