+++ web/src/hooks/index.ts
@@ -25,3 +25,5 @@ export { useSearchParamHelpers } from './useSearchParamHelpers';
 export type { SearchParamUpdates } from './useSearchParamHelpers';
 
 export { usePageKeyboardShortcuts } from './usePageKeyboardShortcuts';
+
+export { useCommandPalette } from './useCommandPalette';
